Changes image dataset data type

+++ rmltraintfbbox/rmltraintfbbox/validation/utils.py
@@ -35,7 +35,7 @@ def get_image_dataset(dir_path, rescale=1.0, gaussian_stddev=0.0):
             dims = tf.cast(tf.shape(img), tf.float32)[:2]
             dims = tf.cast(dims * rescale, tf.int32)
             img = tf.image.resize(img, dims)
-        return tf.cast(img, tf.uint8)
+        return tf.cast(img, tf.float32)
 
     image_files = sorted(glob.glob(os.path.join(dir_path, "image_*")))
     image_dataset = tf.data.Dataset.from_tensor_slices(image_files).map(image_parser, num_parallel_calls=16)
